Fixed regression introduced in 1.1.44 which caused "Fit Image to Screen" to create a tiny window instead of a window which filled the screen.
When unbounding the zoom in commit f93e388, it inadvertently caused calculations in Helpers::GetWindowRectMatchingImageSize() to overflow an int because zoom is capped at DBL_MAX.
In turn, this caused the image zoom to reset to the absolute minimum window size possible, since the dimension were now negative.

Workaround is that when FitToScreen is done, it pretends the image is 65535 on one dimension, and the other dimension is scaled to the ImageAR.
This guarantees that the calculations later will not overflow, and the results revert to pre-1.1.44 behavior.

Closes #158