fix bug where zooming out to 0% and trying to enlarge it gets stuck at 0%
This was due to the possibility that the new X and Y dimensions are still the same (due to rounding) but the zoom is increasing.
The old code would restore the zoom value, which would cause a state that you can't zoom in.
The new code takes into account that, if you're zooming out, it'll restore the old zoom value, but if zooming in, allow the factor to continue to increase, even if the newX and newY don't change.

Closes #159