Asynchronous call to CFileList.
The original behavior of the program is to check all files synchronously upon startup, which in my opinion is a bad practice.
It should however check asynchronously while it's viewing the startup image for quicker startup.
Rationale: I have thousands of images in one folder, and opening each image individually takes about 1-3 seconds to load.
My own solution is immature, but it works.