better fix for the freezing issue #153
Freeze happened in case when conditions passed execution to the thread
pool code. The code used to compute the number of threads to use
expects a slice of at least twice the size of StripPadding, otherwise
it can end up with no threads to use. In case it ended up with no
threads, there was no one to signal the finished event.

As one of the conditions for thread pool processing is to have more
than 100k pixels to process, the necessary conditions couldn't be
triggered for resolutions lower than 3226 px as described in the
original fix, 3226 * 31 = 100006. With standard 4k resolution of 3840px
the hang occurred in nTargetCY range of 27 to 31px.